requirements.txt
Flask==2.0.1
torch==2.3.0
opencv-python-headless==4.5.3.56
Pillow==9.1.0